requirements.txt
absl-py==0.9.0
astor==0.8.1
backports.functools-lru-cache==1.6.1
backports.weakref==1.0.post1
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
configparser==4.0.2
cycler==0.10.0
docker-pycreds==0.4.0
enum34==1.1.10
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.3.0
gast==0.2.2
gitdb2==2.0.6
GitPython==2.1.15
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gql==0.2.0
graphql-core==1.1
grpcio==1.27.2
h5py==2.10.0
idna==2.9
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==2.2.5
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
numpy==1.16.6
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-contrib-python==4.2.0.32
opt-einsum==2.3.2
pathtools==0.1.2
Pillow==6.2.2
promise==2.3
protobuf==3.11.3
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.2.2
sentry-sdk==0.14.2
shortuuid==0.5.0
six==1.13.0
smmap==3.0.1
smmap2==3.0.1
subprocess32==3.5.4
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
typing==3.7.4.1
urllib3==1.25.8
wandb==0.8.30
watchdog==0.10.2
Werkzeug==1.0.0
wrapt==1.12.1
